moving test to pytest.
